+++ server/db/migrate/20260317146080_add_locale_to_spree_orders.spree.rb
@@ -1,6 +1,0 @@
-# This migration comes from spree (originally 20260226000000)
-class AddLocaleToSpreeOrders < ActiveRecord::Migration[7.2]
-  def change
-    add_column :spree_orders, :locale, :string
-  end
-end